+++ explorer/src/lib/chain-info/transformTransaction.js
@@ -2,7 +2,6 @@ import { unixTsToDate } from "$lib/dusk/date";
 
 /** @type {(v: GQLTransaction) => Transaction} */
 const transformTransaction = (tx) => ({
-  amount: undefined,
   blockhash: tx.blockHash,
   blockheight: tx.blockHeight,
   date: unixTsToDate(tx.blockTimestamp),